requirements.txt
Django==3.1.1
psycopg2-binary==2.8.6
python-decouple==3.3
dj_database_url==0.5.0
requests==2.25.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-auth-httplib2==0.0.4
flask==1.1.2
